Adicionando função de cancelar evento para supers admins e admins e filtrar por tipo de evento

+++ codechella/src/main/java/com/example/codechella/models/ingresso/IngressoDTO.java
@@ -2,7 +2,7 @@ package com.example.codechella.models.ingresso;
 
 import java.time.LocalDate;
 
-public record IngressoDTO(Long id, Long eventoId, LocalDate date, int quantidade, int valor, String descricao, TipoStatus tipoStatus) {
+public record IngressoDTO(Long id, Long eventoId, LocalDate date, int quantidade, Double valor, String descricao, TipoStatus tipoStatus) {
 
     public static IngressoDTO toDTO(Ingresso ingresso){
         return new IngressoDTO(ingresso.getId(), ingresso.getEventoId(), ingresso.getDate(), ingresso.getQuantidade(), ingresso.getValor(), ingresso.getDescricao(), ingresso.getStatus());
